move unused code to community project
