+++ src/web/app/pom.xml
@@ -1529,6 +1529,16 @@
       </dependencies>
     </profile>
     <profile>
+      <id>geopkg-output</id>
+      <dependencies>
+        <dependency>
+          <groupId>org.geoserver.extension</groupId>
+          <artifactId>gs-geopkg-output</artifactId>
+          <version>${project.version}</version>
+        </dependency>
+      </dependencies>
+    </profile>
+    <profile>
       <id>mongodb</id>
       <dependencies>
         <dependency>
@@ -1554,6 +1564,7 @@
         </dependency>
       </dependencies>
     </profile>
+
     <profile>
       <id>metadata</id>
       <dependencies>
